Resolve lint issues
